Add files via upload
Multi-layer neural networks will be trained and evaluated based on data from Run or
Walk Reduced dataset. Note that this Run or Walk Final dataset uses only the subset of
data from the original Run or Walk dataset.


Attribute Information:
The dataset comprises the readings of motion sensors recorded while users executed
typical daily activities. The detailed format is described in the package. The attributes
correspond to raw sensor readings. There are a total of 11 attributes.
1. date
2. time
3. username
4. wrist
5. activity
6. acceleration_x
7. acceleration_y
8. acceleration_z
9. gyro_x
10. gyro_y
11. gyro_z